Add `lib.rs` for `Libra`, make it possible to call from outside by `sync` way.
